% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makemcsim.R
\name{makemcsim}
\alias{makemcsim}
\title{Compile \pkg{GNU MCSim} model file}
\usage{
makemcsim(file, init = F)
}
\arguments{
\item{file}{a character of version number.}

\item{init}{a character to assign the installed directory.}
}
\description{
Download the latest or specific version of \pkg{GNU MCSim} from the official website
(\url{https://www.gnu.org/software/mcsim/}) and install it to the system directory.
}
\details{
This function aims to help users install \pkg{GNU MCSim} more easily.
However, if you can not install it through this function.
The additional way is to follow the instruction and install it manually:
\url{https://www.gnu.org/software/mcsim/mcsim.html#Installation}

The default \code{mxstp} is setting to 5000.
The user can increase \code{mxstp} to avoid possible error return.
If you meet any error when conduct sensitivity analysis,
you can this function to reinstall \pkg{GNU MCSim} and set the higher \code{mxstp}.
The default installed \code{directory} is under \code{/home/username} (Linux),
\code{/Users/username} (MacOS),
and \code{C:/Users/} (windows). To install \pkg{GNU MCSim} in Windows, be sure to install Rtools first.
}
\references{
Bois, F. Y., & Maszle, D. R. (1997).
MCSim: a Monte Carlo simulation program.
\emph{Journal of Statistical Software}, 2(9): 1–60.
}
